Move ensure_bytestream to _artifact_utils
