+++ verta/verta/_artifact_utils.py
@@ -1,6 +1,5 @@
 import six
-
-import pickle
+import six.moves.cPickle as pickle
 
 from verta import _utils
 
@@ -8,6 +7,50 @@ try:
     from tensorflow import keras
 except ImportError:
     pass
+
+
+def ensure_bytestream(obj):
+    """
+    Converts an object into a bytestream.
+
+    If `obj` is file-like, its contents will be read into memory and then wrapped in a bytestream.
+    This has a performance cost, but checking beforehand whether an arbitrary file-like object
+    returns bytes is an implementation nightmare.
+
+    If `obj` is not file-like, it will be serialized and then wrapped in a bytestream.
+
+    Parameters
+    ----------
+    obj : file-like or object
+        Object to convert into a bytestream.
+
+    Returns
+    -------
+    file-like
+        Buffered bytestream.
+
+    Raises
+    ------
+    ValueError
+        If `obj` contains no data.
+
+    """
+    if hasattr(obj, 'read'):  # if `obj` is file-like
+        try:  # reset cursor to beginning in case user forgot
+            obj.seek(0)
+        except AttributeError:
+            pass
+        contents = obj.read()  # read to cast into binary
+        try:  # reset cursor to beginning as a courtesy
+            obj.seek(0)
+        except AttributeError:
+            pass
+        if not len(contents):
+            raise ValueError("object contains no data")
+        bytestring = six.ensure_binary(contents)
+    else:  # `obj` is not file-like
+        bytestring = pickle.dumps(obj)
+    return six.BytesIO(bytestring)
 
 
 def serialize_model(model):
@@ -30,7 +73,7 @@ def serialize_model(model):
     try:
         model.save(bytestream)
     except AttributeError:
-        model = _utils.ensure_bytestream(model)
+        model = ensure_bytestream(model)
     else:
         bytestream.seek(0)
         model = bytestream
